backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
pyarrow==14.0.1
openpyxl==3.1.2

# Machine Learning
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# NLP
spacy==3.7.2
langdetect==1.0.9

# File Processing
python-magic==0.4.27
chardet==5.2.0
pdfplumber==0.9.0
python-docx==1.1.0

# Storage
minio==7.2.0
boto3==1.34.0

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Task Queue
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1